requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
Flask-CORS==4.0.0
bcrypt==4.2.0
Pillow==10.4.0
opencv-python==4.10.0.84
face-recognition==1.3.0
dlib-bin==19.24.6
numpy==1.26.4
python-dotenv==1.0.1
geopy==2.4.1
requests==2.32.3
psycopg2-binary
pytesseract==0.3.10
piexif==1.1.3
waitress==3.0.0